requirements.txt
Django==4.1.2
ipython
tzdata
pytz
djangorestframework